Updated Word Processing and Minor Update 1
Added a PDF of planning notes.

Implemented a way to read files for both the Lesson state and the Trivia
state. Created a Java program to help generate Lesson and Trivia
scripts. Included shortcuts to both applications in main directory.

Implemented minor changes to game, such as a timer, a score keeper, and
the beginning of a state machine.

Implemented the beginning of a menu.

Created a Button class in graphics.py.