Audio Feedback for Quiz Mode
Added audio feedback for Quiz mode to alert users to whether or not they got the answer right. The game will not explicitly provide the correct answer to the user, so that the user, so that the user can use this game as a study guide when learning Python.

The sounds in this commit are from the SoundBible.com. The elevator bell sound, by Corsci, is placed  under a Sampling Plus 1.0 license, and has been edited to suit the games needs. A copy of the license can be found here: http://creativecommons.org/licenses/sampling+/1.0/. The Door buzzer sound, by Mike Koenig, is placed under a CC-BY 3.0 license, which can be found here: https://creativecommons.org/licenses/by/3.0/. It has been edited to suit the games needs.